Resolved conflict
