+++ src/main/java/com/kodilla/ecommercee/domain/CartItem.java
@@ -1,0 +1,2 @@
+package com.kodilla.ecommercee.domain;public class CartItem {
+}
