--- src/main/java/com/kodilla/ecommercee/domain/OrderDto.java
@@ -1,17 +1,0 @@
-package com.kodilla.ecommercee.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-import java.util.Optional;
-
-@Getter
-@AllArgsConstructor
-public class OrderDto {
-
-    private Long id;
-    private Long userId;
-    private Optional<LocalDateTime> dateTime;
-    private CartStatus cartStatus;
-}
