+++ src/main/java/com/kodilla/ecommercee/domain/dto/CartItemDto.java
@@ -1,15 +1,17 @@
 package com.kodilla.ecommercee.domain.dto;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
 public class CartItemDto {
-
     private long id;
-    private long product_id;
-    private long order_id;
+    private long productId;
     private long quantity;
-    private long unit_Price;
+    private BigDecimal unitPrice;
+    private long orderId;
 }
